% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalHelpers.R
\name{check_sequence}
\alias{check_sequence}
\title{Sequence Checking}
\usage{
check_sequence(t = NULL, n)
}
\arguments{
\item{t}{A numeric vector of length n.}

\item{n}{A atomic numeric.}
}
\value{
A numeric vector.
}
\description{
Sequence Checking
}
