% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mankendall.R
\name{mk_test}
\alias{mk_test}
\title{Mann-Kendall Test}
\usage{
mk_test(
  t = NULL,
  Y,
  alpha_val = 0.05,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{t}{A vector of ranks (numeric, or coercible to numeric).
For Time Series the observation dates.}

\item{Y}{A vector of ranks. For Time Series the values observed at each
\emph{t}.}

\item{alpha_val}{Level of significance used for testing.}

\item{alternative}{What should \mjseqn{H_{1}} be?}
}
\value{
Built in tests like the \code{t.test} use a class "htest".
Hopefully this will be my return value/class as well.
}
\description{
Mann-Kendall Test
}
\seealso{
\code{\link{mk_score}} \code{\link{mk_variance}}
\code{\link{mk_tau}}
}
