% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mankendall.R
\name{mk_variance}
\alias{mk_variance}
\title{Calculate Variance of Mann-Kendall Test Statistic}
\usage{
mk_variance(n, t = NULL, u = NULL)
}
\arguments{
\item{n}{Length of ranking/time series.}

\item{t}{Contingency table of ranking \mjseqn{R_{t}}. If \code{NULL} (the default),
assumes that no duplications are present.}

\item{u}{Contingency table of ranking \mjseqn{R_{u}}. If \code{NULL} (the default),
assumes that no duplications are present.}
}
\value{
Variance of Kendall's test statistic
}
\description{
Calculate Variance of Mann-Kendall Test Statistic
}
\details{
\loadmathjax \mjsdeqn{\sigma_S^2 = \frac{(n(n - 1)(2n + 5) -
\sum_{t \in g_t}t(t - 1)(2t + 5) - \sum_{u \in g_u}
u(u - 1)(2u + 5)) + \frac{1}{9n(n-1)(n-2)} \times
\sum_{t \in g_t} t(t-1)(t-2) \times \sum_{u \in g_u} u(u-1)(u-2) +
\frac{1}{2n(n-1)} \times \sum_{t \in g_t} t(t-1) \times
\sum_{u \in g_u} u(u-1)}{18}}
}
\seealso{
\code{\link{mk_test}}
}
