% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalHelpers.R
\name{onesided_sgn}
\alias{onesided_sgn}
\title{Calculate sum of signs of one vector}
\usage{
onesided_sgn(t)
}
\arguments{
\item{t}{A vector of ranks.}
}
\value{
An Integer.
}
\description{
This fuction takes a vector of ranks as
input and returns the summed product of
the signs of \eqn{\binom {n} {2}} permutations.
}
\details{
\code{onesided_sgn} operates like \code{\link{sgn}} but takes only
a single vector as input. In the context of the Kendall's Rank Correlation
Coefficient this assumes, that the provided vector represents entries in
their natural order.
}
\references{
Kendall, M. G. (1975): Rank Correlation Methods. 4th Edition London, High Wycombe: Charles Griffin & Company Ltd.
}
\seealso{
\code{\link{sgn}}
}
