% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalHelpers.R
\name{sgn}
\alias{sgn}
\title{Calculate sum of signs of two vectors}
\usage{
sgn(t, z)
}
\arguments{
\item{t}{A vector of ranks}

\item{z}{A vector of ranks}
}
\value{
An Integer
}
\description{
Der folgende Satz stimmt zwar aktuell, Funktionsweise soll
aber geändert werden
This function takes two matrices as input (generated by combn) and returns
the summed product of the form \code{sign(z2 - z1) * sign(t2 - t2)}.
}
\references{
Kendall der olle Hengst
}
\seealso{
\code{\link{onesided_sgn}}
}
