% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theilsen.R
\name{ts_score}
\alias{ts_score}
\title{Calculate Test Statistic for Sen-Slope}
\usage{
ts_score(t = NULL, Y, slope)
}
\arguments{
\item{t}{A vector of ranks (numeric, or coercible to numeric).
For Time Series the observation dates.}

\item{Y}{A vector of ranks (numeric, or coercible to numeric).
For Time Series the observation dates.}

\item{slope}{Slope of Rankings.}
}
\description{
Calculate Test Statistic of Sen's Slope.
}
\seealso{
\code{\link{ts_test}}
}
