% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theilsen.R
\name{ts_slope}
\alias{ts_slope}
\title{Calculate Sen-Slope}
\usage{
ts_slope(t = NULL, Y)
}
\arguments{
\item{t}{A vector of ranks (numeric, or coercible to numeric).
For Time Series the observation dates.}

\item{Y}{A vector of ranks. For Time Series the values observed at each
\emph{t}.}
}
\value{

}
\description{
\loadmathjax Calculation of Sen's slope. See Details.
}
\details{
Sen's slope is calculated as followed:
\mjsdeqn{b = \frac{x_{n/2} + x_{n/2 + 1}}{2}}
with \eqn{x} being the differences of every \mjseqn{\binom{n}{2}} value pairs:
\mjsdeqn{x = (Y_{j} - Y_{i})/(t_{j} - t_{i}), 1 \le \, i < \, j \, \le n}
}
\references{
Sen, Pranab Kumar (1968): Estimates of the Regression
Coefficient Based on Kendall's Tau. In: Journal of the American
Statistical Association 63 (324), P. 1379-1389. DOI: 10.2307/2285891
}
\seealso{
\code{\link{ts_test}}
}
