% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theilsen.R
\name{ts_tau}
\alias{ts_tau}
\title{Calculate Theil-Sen Equivalent of Kendall's Tau}
\usage{
ts_tau(t = NULL, n, score)
}
\arguments{
\item{t}{}

\item{n}{}

\item{score}{}
}
\value{

}
\description{
Sen denotes this as \emph{U}. It's
functionally equivalent to Kendall's Tau,
however their calculations differ.
}
\seealso{
\code{\link{ts_test}}
}
