% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theilsen.R
\name{ts_variance}
\alias{ts_variance}
\title{Calculate Variance of Sen's Test Statistic}
\usage{
ts_variance(n, u = rep(1, length.out = n))
}
\arguments{
\item{n}{A numeric. Length of the ranking.}

\item{u}{Contingency table of ranks in one ranking.
Will be coerced to numeric vector.}
}
\value{

}
\description{
Calculates variance Sen's Test Statistic Z.
}
\details{
This is equivalent to the calculation of the
variance Kendall's test statistic, accounting for
duplicates in only one of the rankings:
\deqn{\sigma^{2}_{S} = \frac{n(n-1)(2n+5) \times \sum_{t \in g_{e}}t(t-1)(2t+5)}{18}}
}
\seealso{
\code{\link{ts_test}}
}
