% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mankendall.R
\name{mk_tau}
\alias{mk_tau}
\title{Calculate Kendall's Tau}
\usage{
mk_tau(score, n, t = NULL, u = NULL)
}
\arguments{
\item{score}{Man Kendall Score, see \code{\link{mk_score}}.}

\item{n}{Length of ranking/time series.}

\item{t}{Contingency table of ranking \eqn{R_{t}}. If \code{NULL} (the default),
assumes that no duplications are present.}

\item{u}{Contingency table of ranking \eqn{R_{u}}. If \code{NULL} (the default),
assumes that no duplications are present.}
}
\value{
A named vector (Tau a or Tau b, depending on the input).
}
\description{
Calculate Kendall's Tau
}
