% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mankendall.R
\name{mk_test}
\alias{mk_test}
\title{Mann-Kendall Test}
\usage{
mk_test(
  t = NULL,
  Y,
  alpha_val = 0.05,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{t}{}

\item{Y}{}

\item{alpha_val}{}

\item{alternative}{}
}
\value{
Built in tests like the \code{t.test} use a class "htest".
Hopefully this will be my return value/class as well.
}
\description{
Mann-Kendall Test
}
