% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalHelpers.R
\name{sgn}
\alias{sgn}
\title{Calculate sum of signs of two vectors}
\usage{
sgn(t, z)
}
\arguments{
\item{t}{A vector of ranks.}

\item{z}{A vector of ranks.}
}
\value{
An Integer.
}
\description{
Takes two vectors as input and returns summed product of the form
\eqn{\textrm{sign}(z_j - z_i) \* \textrm{sign}(t_j - t_i)}.
}
\references{
Kendall, M. G. (1975): Rank Correlation Methods. 4th Edition London,
High Wycombe: Charles Griffin & Company Ltd.
}
\seealso{
\code{\link{onesided_sgn}}
}
