% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theilsen.R
\name{ts_score}
\alias{ts_score}
\title{Calculate Test Statistic for Sen-Slope}
\usage{
ts_score(t = NULL, Y, slope)
}
\arguments{
\item{t}{}

\item{Y}{}

\item{slope}{}
}
\value{

}
\description{
Calculate Test Statistic of Sen's Slope.
}
\details{
This
}
\seealso{
\code{\link{ts_test}}
}
