% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theilsen.R
\name{ts_test}
\alias{ts_test}
\title{Test Sen-Slope}
\usage{
ts_test(t = NULL, Y, alpha_val = 0.05)
}
\arguments{
\item{t}{A vector of ranks (numeric, or coercible to numeric).
For Time Series the observation dates.}

\item{Y}{A vector of ranks. For Time Series the values observed at each
\emph{t}.}

\item{alpha_val}{Level of significance used for testing.}
}
\value{
Built in tests like the \code{t.test} use a class "htest".
Hopefully this will be my return value/class as well.
}
\description{
Tests if Sen's slope is significant. It's effectively
a wrapper around all functions in this package starting
with \code{ts_}.
}
\seealso{
\code{\link{ts_slope}} \code{\link{ts_variance}}
\code{\link{ts_tau}} \code{\link{ts_score}}
}
